Tests: moved BasePair related tests to Tests/test_basepair.py
