+++ Tests/Makefile
@@ -3,8 +3,9 @@ unittests: testcpp testpython
 testpython:
 	echo "running python tests"
 	python unittest1.py
-	python test_movement.py
 	python test_dna.py
+	python test_basepair.py
+	python test_movement.py
 	cd heligeom && python heligeom_test.py -v
 
 testcpp:
